First part of moving the dropdown positioning to own file
