+++ src/js/language-toast.js
@@ -1,41 +1,7 @@
+import { positionDropdown } from './position-dropdown';
+
 document.addEventListener('DOMContentLoaded', function languageToastPositioning() {
   const buttons = document.querySelectorAll('.nav-toggle__button.has-toast');
-
-  function positionToast(button, toast) {
-    if (!toast || !toast.classList.contains('nav-toggle-dropdown--language-toast')) {
-      return;
-    }
-
-    // Get button and toast dimensions and position relative to the viewport.
-    const rect = button.getBoundingClientRect();
-    const buttonWidth = button.offsetWidth;
-    const toastWidth = toast.offsetWidth;
-    const viewportWidth = document.documentElement.clientWidth;
-    const gutter = 16;
-
-    // Positioning if the toast doesn't fit on the right side.
-    if ((viewportWidth - rect.right) < ((toastWidth / 2) - (buttonWidth / 2) + gutter)) {
-      const right = viewportWidth - rect.right - gutter;
-      const toastArrowRight = ((viewportWidth - rect.right) - gutter) + (buttonWidth / 2);
-
-      toast.style.left = 'auto';
-      toast.style.transform = 'none';
-      toast.style.right = `-${right}px`;
-      toast.style.setProperty('--toast-arrow-left', 'auto');
-      toast.style.setProperty('--toast-arrow-right', `${toastArrowRight}px`);
-      toast.style.setProperty('--toast-arrow-transform', '50%');
-    }
-
-    // Positioning if the toast doesn't fit on the left side.
-    if ((rect.left) < ((toastWidth / 2) - (buttonWidth / 2) + gutter)) {
-      const left = rect.left - gutter;
-      const toastArrowLeft = (rect.left - gutter) + (buttonWidth / 2);
-
-      toast.style.left = `-${left}px`;
-      toast.style.transform = 'none';
-      toast.style.setProperty('--toast-arrow-left', `${toastArrowLeft}px`);
-    }
-  }
 
   // Handle click event and position the toast correctly before opening it.
   buttons.forEach(button => {
@@ -51,7 +17,7 @@ document.addEventListener('DOMContentLoaded', function languageToastPositioning(
       toast.removeAttribute('style');
 
       // Position and after that show the toast.
-      positionToast(button, toast);
+      positionDropdown(button);
     });
   });
 
@@ -63,7 +29,7 @@ document.addEventListener('DOMContentLoaded', function languageToastPositioning(
 
       if (toast && !toast.classList.contains('nav-toggle-dropdown--closed')) {
         toast.removeAttribute('style');
-        positionToast(button, toast);
+        positionDropdown(button);
       }
     });
   });
